fixed f8e130c from https://gitee.com/yuanmmy/communication_dsoftbus_5/pulls/2735 codex bugfix Signed-off-by: y30033032 <yuanmeng36@huawei.com>

+++ tests/sdk/bus_center/fuzztest/shiftlnngear_fuzzer/shiftlnngear_fuzzer.cpp
@@ -63,7 +63,7 @@ namespace OHOS {
                 break;
         }
         g_mode.wakeupFlag = (size % MAX_SIZE_WAKEUP_MODE) ? true : false;
-        int32_t callerIdLen = size % CALLER_ID_MAX_LEN + 1;
+        size_t callerIdLen = size % CALLER_ID_MAX_LEN + 1;
         callerId = (char *)malloc(callerIdLen);
         if (callerId == nullptr) {
             return;
